feat: init page directory, actions bar component
