+++ frontend/src/pages/recipes.tsx
@@ -1,5 +1,5 @@
 import Layout from '@/components/layout/Layout';
-import AllRecipesView from '@/components/AllRecipesView';
+import AllRecipesView from '@/components/pages/recipes/AllRecipesView';
 
 export default function Recipes() {
   return (
